Update base for Update on "[wip][inductor]lowering scan to while_loop"
cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy yf225 chenyang78 kadeng muchulee8 amjames chauhang aakhundov

[ghstack-poisoned]

+++ test/inductor/test_aot_inductor.py
@@ -1586,7 +1586,7 @@ class AOTInductorTestsTemplate:
 
     @common_utils.parametrize("dynamic", [False, True])
     def test_while_loop_with_conv(self, dynamic):
-        inputs = (torch.randn(2, 4, 4, 4, device=self.device),)
+        inputs = (torch.randn(2, 4, 4, 4, device=self.device, dtype=torch.float64),)
         dim0_ab = Dim("s0", min=2, max=1024)
         dynamic_shapes = None
         if dynamic:
